requirements.txt
streamlit
requests
pandas
st-gsheets-connection
Pillow